refactor(JobSections): break down 130-line component file into individual section modules (→11 lines)
- Extract DescriptionSection (14 lines): role description with HTML formatting
- Extract ResponsibilitiesSection (18 lines): key responsibilities list
- Extract RequirementsSection (18 lines): job requirements list
- Extract SkillsSection (30 lines): skills grouped by category
- Extract ExperienceSection (9 lines): experience level display
- Extract BenefitsSection (20 lines): benefits grid with icons
- Extract CultureSection (9 lines): company culture description
- Extract AdditionalSection (9 lines): additional info display
- Create JobSections/index.js for clean re-exports
- Main JobSections.jsx reduced from 130→11 lines (re-export only)
- Each section independently testable and maintainable
- Follows established feature module pattern

Part of #203 (files over 200 lines reduction)
