+++ apps/registry/app/jobs/JobDetailModule/components/JobSections/AdditionalSection.jsx
@@ -1,0 +1,9 @@
+export const AdditionalSection = ({ additional }) => {
+  if (!additional) return null;
+  return (
+    <div>
+      <h2 className="text-xl font-semibold mb-4">Additional Information</h2>
+      <p className="text-gray-600">{additional}</p>
+    </div>
+  );
+};
