+++ apps/registry/app/jobs/JobDetailModule/components/JobSections/index.js
@@ -1,0 +1,8 @@
+export { DescriptionSection } from './DescriptionSection';
+export { ResponsibilitiesSection } from './ResponsibilitiesSection';
+export { RequirementsSection } from './RequirementsSection';
+export { SkillsSection } from './SkillsSection';
+export { ExperienceSection } from './ExperienceSection';
+export { BenefitsSection } from './BenefitsSection';
+export { CultureSection } from './CultureSection';
+export { AdditionalSection } from './AdditionalSection';
